Changed response codes for API/Delete

+++ src/ApiBundle/Application/Posts.php
@@ -5,6 +5,7 @@ namespace ApiBundle\Application;
 use BlogBundle\Entity\Post;
 use JMS\Serializer\Serializer;
 use BlogBundle\Domain\BasicInterface as BasicDomainInterface;
+use Symfony\Component\HttpKernel\Exception\GoneHttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class Posts extends AbstractApplication
@@ -29,13 +30,10 @@ class Posts extends AbstractApplication
      */
     public function removeById($id)
     {
-        try {
-            $service = $this->getService();
-            $entity = $service->loadById($id);
-            $service->remove($entity);
-        } catch (\Exception $e) {
-            // Do nothing
-        }
+        $service = $this->getService();
+        $entity = $service->loadById($id);
+        if (empty($entity)) throw new GoneHttpException("Post with id: '{$id}' already deleted");
+        $service->remove($entity);
     }
 
     /**
